pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "be-system"
version = "0.1.0"
description = "MVP multi-agent system for synthetic bioequivalence study design"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "biopython",
  "openai",
  "pydantic",
  "python-dotenv",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
